Reuse request body with net/http.Request.GetBody if available
Commit 3b7925a addressed the problem where Transport.RoundTrip didn't
work with a non-nil body since the original io.Reader had been emptied
by the first request (i.e. the second request got an immediate EOF) but
did so by reading the whole payload into RAM. That's totally fine for
small requests but didn't scale with large payloads and/or high
concurrency.

We address this by paying attention to net/http.Request.GetBody
(introduced in Go 1.8), which if non-nil contains the address of a
function that produces a fresh io.Reader. If that field is unset we
fall back to the old behavior of reading the payload to RAM.